Develop

+++ controllers/contactControllers.js
@@ -3,15 +3,15 @@ const Contact = require('../models/contactModel')
 
 //@desc Get All Contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getAllContacts = asyncHandler(async (req, res) => {
-  const allContacts = await Contact.find()
-  res.status(200).json(allContacts)
+  const contacts = await Contact.find({ user_id: req.user.id })
+  res.status(200).json(contacts)
 })
 
 //@desc Create Contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
   const { name, email, phone } = req.body
 
@@ -19,10 +19,12 @@ const createContact = asyncHandler(async (req, res) => {
     res.status(400)
     throw new Error('All fields are mandatory!')
   }
+
   const contact = await Contact.create({
     name,
     email,
     phone,
+    user_id: req.user.id,
   })
 
   res.status(201).json(contact)
@@ -30,7 +32,7 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc Get Single Contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
   const { id } = req.params
   const contact = await Contact.findById(id)
@@ -44,7 +46,7 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Update Contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
   const { id } = req.params
   const contact = await Contact.findById(id)
@@ -52,6 +54,13 @@ const updateContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404)
     throw new Error('Contact was not found!')
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403)
+    throw new Error(
+      "The User doesn't have permission to update another user's contact!"
+    )
   }
   const updatedContact = await Contact.findByIdAndUpdate(id, req.body, {
     new: true,
@@ -61,7 +70,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete Contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
   const { id } = req.params
   const contact = await Contact.findById(id)
@@ -69,6 +78,13 @@ const deleteContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404)
     throw new Error('Contact was not found!')
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403)
+    throw new Error(
+      "The User doesn't have permission to delete another user's contact!"
+    )
   }
   const deletedContact = await Contact.findByIdAndDelete(id)
   res.status(200).json(deletedContact)
